+++ lisa/src/test/java/it/unive/lisa/imp/IMPFrontend.java
@@ -21,17 +21,6 @@ import it.unive.lisa.program.CompilationUnit;
 import it.unive.lisa.program.Global;
 import it.unive.lisa.program.Program;
 import it.unive.lisa.program.SourceCodeLocation;
-import it.unive.lisa.program.annotations.Annotation;
-import it.unive.lisa.program.annotations.AnnotationMember;
-import it.unive.lisa.program.annotations.Annotations;
-import it.unive.lisa.program.annotations.values.AnnotationValue;
-import it.unive.lisa.program.annotations.values.ArrayAnnotationValue;
-import it.unive.lisa.program.annotations.values.BasicAnnotationValue;
-import it.unive.lisa.program.annotations.values.BoolAnnotationValue;
-import it.unive.lisa.program.annotations.values.CompilationUnitAnnotationValue;
-import it.unive.lisa.program.annotations.values.FloatAnnotationValue;
-import it.unive.lisa.program.annotations.values.IntAnnotationValue;
-import it.unive.lisa.program.annotations.values.StringAnnotationValue;
 import it.unive.lisa.program.cfg.CFG;
 import it.unive.lisa.program.cfg.CFGDescriptor;
 import it.unive.lisa.program.cfg.Parameter;
@@ -39,13 +28,6 @@ import it.unive.lisa.program.cfg.statement.Statement;
 import it.unive.lisa.program.cfg.statement.UnresolvedCall.ResolutionStrategy;
 import it.unive.lisa.test.antlr.IMPLexer;
 import it.unive.lisa.test.antlr.IMPParser;
-import it.unive.lisa.test.antlr.IMPParser.AnnotationContext;
-import it.unive.lisa.test.antlr.IMPParser.AnnotationMemberContext;
-import it.unive.lisa.test.antlr.IMPParser.AnnotationMembersContext;
-import it.unive.lisa.test.antlr.IMPParser.AnnotationValueContext;
-import it.unive.lisa.test.antlr.IMPParser.AnnotationsContext;
-import it.unive.lisa.test.antlr.IMPParser.ArrayAnnotationValueContext;
-import it.unive.lisa.test.antlr.IMPParser.BasicAnnotationValueContext;
 import it.unive.lisa.test.antlr.IMPParser.ConstructorDeclarationContext;
 import it.unive.lisa.test.antlr.IMPParser.FieldDeclarationContext;
 import it.unive.lisa.test.antlr.IMPParser.FileContext;
@@ -61,9 +43,7 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import org.antlr.v4.runtime.BailErrorStrategy;
 import org.antlr.v4.runtime.CharStreams;
@@ -263,7 +243,7 @@ public class IMPFrontend extends IMPParserBaseVisitor<Object> {
 	@Override
 	public Global visitFieldDeclaration(FieldDeclarationContext ctx) {
 		return new Global(new SourceCodeLocation(file, getLine(ctx), getCol(ctx)), ctx.name.getText(),
-				Untyped.INSTANCE, visitAnnotations(ctx.annotations()));
+				Untyped.INSTANCE, new IMPAnnotationVisitor().visitAnnotations(ctx.annotations()));
 	}
 
 	@Override
@@ -283,7 +263,8 @@ public class IMPFrontend extends IMPParserBaseVisitor<Object> {
 	private CFGDescriptor mkDescriptor(ConstructorDeclarationContext ctx) {
 		CFGDescriptor descriptor = new CFGDescriptor(new SourceCodeLocation(file, getLine(ctx), getCol(ctx)),
 				currentUnit,
-				true, ctx.name.getText(), Untyped.INSTANCE, visitAnnotations(ctx.annotations()),
+				true, ctx.name.getText(), Untyped.INSTANCE,
+				new IMPAnnotationVisitor().visitAnnotations(ctx.annotations()),
 				visitFormals(ctx.formals()));
 		descriptor.setOverridable(false);
 		return descriptor;
@@ -292,7 +273,8 @@ public class IMPFrontend extends IMPParserBaseVisitor<Object> {
 	private CFGDescriptor mkDescriptor(MethodDeclarationContext ctx) {
 		CFGDescriptor descriptor = new CFGDescriptor(new SourceCodeLocation(file, getLine(ctx), getCol(ctx)),
 				currentUnit,
-				true, ctx.name.getText(), Untyped.INSTANCE, visitAnnotations(ctx.annotations()),
+				true, ctx.name.getText(), Untyped.INSTANCE,
+				new IMPAnnotationVisitor().visitAnnotations(ctx.annotations()),
 				visitFormals(ctx.formals()));
 
 		if (ctx.FINAL() != null)
@@ -317,82 +299,6 @@ public class IMPFrontend extends IMPParserBaseVisitor<Object> {
 	@Override
 	public Parameter visitFormal(FormalContext ctx) {
 		return new Parameter(new SourceCodeLocation(file, getLine(ctx), getCol(ctx)), ctx.name.getText(),
-				Untyped.INSTANCE, visitAnnotations(ctx.annotations()));
-	}
-
-	@Override
-	public Annotations visitAnnotations(AnnotationsContext ctx) {
-		if (ctx == null)
-			return new Annotations();
-		List<Annotation> anns = new ArrayList<>();
-		for (int i = 0; i < ctx.annotation().size(); i++)
-			anns.add(visitAnnotation(ctx.annotation(i)));
-		return new Annotations(anns);
-	}
-
-	@Override
-	public Annotation visitAnnotation(AnnotationContext ctx) {
-		String annotationName = ctx.IDENTIFIER().getText();
-		if (annotationName.startsWith("Inherited"))
-			return new Annotation(annotationName, visitAnnotationMembers(ctx.annotationMembers()), true);
-		else
-			return new Annotation(annotationName, visitAnnotationMembers(ctx.annotationMembers()), false);
-	}
-
-	@Override
-	public List<AnnotationMember> visitAnnotationMembers(AnnotationMembersContext ctx) {
-		List<AnnotationMember> arr = new ArrayList<>();
-		if (ctx == null)
-			return arr;
-		else
-			for (int i = 0; i < ctx.annotationMember().size(); i++)
-				arr.add(visitAnnotationMember(ctx.annotationMember(i)));
-		return arr;
-	}
-
-	@Override
-	public AnnotationMember visitAnnotationMember(AnnotationMemberContext ctx) {
-		return new AnnotationMember(ctx.IDENTIFIER().getText(), visitAnnotationValue(ctx.annotationValue()));
-	}
-
-	@Override
-	public AnnotationValue visitAnnotationValue(AnnotationValueContext ctx) {
-		if (ctx.basicAnnotationValue() != null)
-			return visitBasicAnnotationValue(ctx.basicAnnotationValue());
-		else
-			return visitArrayAnnotationValue(ctx.arrayAnnotationValue());
-	}
-
-	@Override
-	public BasicAnnotationValue visitBasicAnnotationValue(BasicAnnotationValueContext ctx) {
-		if (ctx.LITERAL_DECIMAL() != null)
-			if (ctx.SUB() != null)
-				return new IntAnnotationValue(Integer.parseInt(ctx.LITERAL_DECIMAL().getText()));
-			else
-				return new IntAnnotationValue(-Integer.parseInt(ctx.LITERAL_DECIMAL().getText()));
-		else if (ctx.LITERAL_FLOAT() != null)
-			if (ctx.SUB() != null)
-				return new FloatAnnotationValue(Float.parseFloat(ctx.LITERAL_FLOAT().getText()));
-			else
-				return new FloatAnnotationValue(-Float.parseFloat(ctx.LITERAL_FLOAT().getText()));
-		else if (ctx.LITERAL_BOOL() != null)
-			return new BoolAnnotationValue(Boolean.parseBoolean(ctx.LITERAL_BOOL().getText()));
-		else if (ctx.LITERAL_STRING() != null)
-			return new StringAnnotationValue(ctx.LITERAL_STRING().getText());
-		else if (ctx.unitName != null)
-			return new CompilationUnitAnnotationValue(ctx.unitName.getText());
-		throw new UnsupportedOperationException("Annotation value not supported: " + ctx);
-	}
-
-	@Override
-	public ArrayAnnotationValue visitArrayAnnotationValue(ArrayAnnotationValueContext ctx) {
-		if (ctx.basicAnnotationValue() == null)
-			return new ArrayAnnotationValue(new BasicAnnotationValue[0]);
-		else {
-			BasicAnnotationValue[] arr = new BasicAnnotationValue[ctx.basicAnnotationValue().size()];
-			for (int i = 0; i < ctx.basicAnnotationValue().size(); i++)
-				arr[i] = visitBasicAnnotationValue(ctx.basicAnnotationValue(i));
-			return new ArrayAnnotationValue(arr);
-		}
+				Untyped.INSTANCE, new IMPAnnotationVisitor().visitAnnotations(ctx.annotations()));
 	}
 }
